splitting configuration from authorization
